renaming of rate files
